requirements.txt
djangorestframework>=3.13.1,<3.14.0
Django>=3.2.5,<4.0
flake8>=4.0.1,<4.1.0
psycopg2>=2.9.3,<2.10.0
Pillow>=9.0.0,<9.1.0
django-cors-headers>=3.11.0,<3.12.0